+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.10.0] - 2017-03-30
+### Added
+- Added the library `fisharebest/localization` which handles the retrieval of the output of the localized name of a language pack and the script direction
+- Added the new trait `AvailableDesignsTrait` which makes it possible to share the design retrieval logic between the installer and the rest of the system
+- Made it possible to use the AJAX-form hash change logic for redirects, too
+- Added the possibility to set a reply-to address or sender address to the `Mailer` class
+
+### Changed
+- Added some guards to various modules to make them more robust 
+- Replaced the library `umpirsky/country-list` with `giggsey/locale` because of its significantly reduced storage footprint 
+
 ## [4.9.2] - 2017-03-26
 ### Fixed
 - Fixed the `uri` Smarty function
@@ -326,7 +337,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - It should be possible again to save ACL roles
 - Silenced a possible PHP warning when trying to login with incorrect credentials 
 
-[Unreleased]: https://github.com/ACP3/cms/compare/v4.9.2...HEAD
+[Unreleased]: https://github.com/ACP3/cms/compare/v4.10.0...HEAD
+[4.10.0]: https://github.com/ACP3/cms/compare/v4.9.2...v4.10.0
 [4.9.2]: https://github.com/ACP3/cms/compare/v4.9.1...v4.9.2
 [4.9.1]: https://github.com/ACP3/cms/compare/v4.9.0...v4.9.1
 [4.9.0]: https://github.com/ACP3/cms/compare/v4.8.5...v4.9.0